requirements.txt
django-pyhub-rag
django-environ
django-debug-toolbar
django-extensions
django-lifecycle
openai
anthropic

# sqlite
sqlite-vec
numpy

# postgres
# psycopg2-binary
# pgvector

ipython
